Add unit test
As the first false positive I wanted to verify was gone now that we switched
to NVD version matching, https://github.com/Nix-Security-WG/nix-security-tracker/issues/43,
was actually just gone because the CVE couldn't be parsed in the first place.

Not working yet, struggling with hspec/LogT/ReaderT syntax
