+++ CVENix.cabal
@@ -50,6 +50,18 @@ library
     default-language: Haskell2010
     ghc-options: -threaded -O2
 
+test-suite Tests
+    type: exitcode-stdio-1.0
+    main-is: Main.hs
+    hs-source-dirs: tests
+    build-depends:
+        base ^>=4.17.2.0,
+        hspec,
+        QuickCheck,
+        HUnit,
+        CVENix,
+        aeson,
+
 executable CVENix
     import:           warnings
     main-is:          Main.hs
