+++ tests/resources/CVE-2023-32611.json
@@ -1,0 +1,1 @@
+{"id":"CVE-2023-32611","sourceIdentifier":"secalert@redhat.com","vulnStatus":"Modified","published":"2023-09-14T20:15:09.550","lastModified":"2023-11-27T14:15:07.713","evaluatorComment":null,"evaluatorSolution":null,"evaluatorImpact":null,"cisaExploitAdd":null,"cisaActionDue":null,"cisaRequiredAction":null,"cisaVulnerabilityName":null,"descriptions":[{"lang":"en","value":"A flaw was found in GLib. GVariant deserialization is vulnerable to a slowdown issue where a crafted GVariant can cause excessive processing, leading to denial of service."},{"lang":"es","value":"Se encontró una falla en GLib. La deserialización de GVariant es vulnerable a un problema de desaceleración en el que un GVariant manipulado puede provocar un procesamiento excesivo y provocar una denegación de servicio."}],"references":[{"url":"https://access.redhat.com/security/cve/CVE-2023-32611","source":"secalert@redhat.com","tags":["Third Party Advisory"]},{"url":"https://bugzilla.redhat.com/show_bug.cgi?id=2211829","source":"secalert@redhat.com","tags":["Issue Tracking","Third Party Advisory"]},{"url":"https://gitlab.gnome.org/GNOME/glib/-/issues/2797","source":"secalert@redhat.com","tags":["Vendor Advisory"]},{"url":"https://lists.debian.org/debian-lts-announce/2023/09/msg00030.html","source":"secalert@redhat.com","tags":null},{"url":"https://security.gentoo.org/glsa/202311-18","source":"secalert@redhat.com","tags":null},{"url":"https://security.netapp.com/advisory/ntap-20231027-0005/","source":"secalert@redhat.com","tags":null}],"metrics":{"cvssMetricV31":[{"cvssData":{"attackComplexity":"LOW","attackVector":"LOCAL","availabilityImpact":"HIGH","baseScore":5.5,"baseSeverity":"MEDIUM","confidentialityImpact":"NONE","integrityImpact":"NONE","privilegesRequired":"NONE","scope":"UNCHANGED","userInteraction":"REQUIRED","vectorString":"CVSS:3.1/AV:L/AC:L/PR:N/UI:R/S:U/C:N/I:N/A:H","version":"3.1"},"exploitabilityScore":1.8,"impactScore":3.6,"source":"nvd@nist.gov","type":"Primary"},{"cvssData":{"attackComplexity":"LOW","attackVector":"LOCAL","availabilityImpact":"HIGH","baseScore":5.5,"baseSeverity":"MEDIUM","confidentialityImpact":"NONE","integrityImpact":"NONE","privilegesRequired":"NONE","scope":"UNCHANGED","userInteraction":"REQUIRED","vectorString":"CVSS:3.1/AV:L/AC:L/PR:N/UI:R/S:U/C:N/I:N/A:H","version":"3.1"},"exploitabilityScore":1.8,"impactScore":3.6,"source":"secalert@redhat.com","type":"Secondary"}]},"weaknesses":[{"source":"nvd@nist.gov","type":"Primary","description":[{"lang":"en","value":"CWE-400"}]},{"source":"secalert@redhat.com","type":"Secondary","description":[{"lang":"en","value":"CWE-400"}]}],"configurations":[{"operator":null,"negate":null,"nodes":[{"operator":"OR","negate":false,"cpeMatch":[{"vulnerable":true,"criteria":"cpe:2.3:a:gnome:glib:*:*:*:*:*:*:*:*","matchCriteriaId":"75A30FA8-F2FB-4637-9EBB-5F2F8D057180","versionStartExcluding":null,"versionStartIncluding":null,"versionEndExcluding":"2.74.2","versionEndIncluding":null}]}]}],"vendorComments":null}
